app/requirements.txt
certifi==2020.6.20
chardet==3.0.4
chromedriver==2.24.1
idna==2.10
lxml==4.5.2
PyYAML==5.3.1
requests==2.24.0
requests-file==1.5.1
selenium==3.141.0
selenium-requests==1.3
simple-term-menu==0.10.1
six==1.15.0
tldextract==2.2.3
tqdm==4.49.0
urllib3==1.25.10
